requirements.txt
google-generativeai>=0.3.0
pillow>=9.0.0
streamlit>=1.28.0
requests>=2.28.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0